Refactor code base and enhance scrolling functionality
Refactored the codebase, renaming all 'clientsStore' references to 'clientStore' in various files. Improved the checkIfScrollToBottomIsNeeded function by considering the sender's clientId when checking whether to scroll to the bottom of the chat. Also enhanced the TypingIndicator component to better handle hover events. Lastly, made minor tweaks to the styles and functionality of several other components.